Don't render fixed-width as fixed-width in tables
Adds context to HipChat `process_directive` calls (just like for Slack)
so we can know to not render fixed-width code when it's inside a
fixed-width block, like our tables.

(We already had context in the Slack renderer, so we
do the same thing there.)

HipChat will remove `<code>` tags that appear inside `<pre>` blocks,
which can throws off the padding in our rendered tables. In Slack,
without this code, we'd see the tick-marks, which isn't a huge deal
overall, but is annoying.

Part of operable/cog#1239