Include newline tokens in tag parsing rules
By tokenizing newlines and including them in the tag reduction rules, we
can be smarter about how we construct tag bodies. When tags start and
end on a newline, we will now strip the newlines surrounding the tag
body and pull the trailing newline into the tag body, making rendering
lists with each tags and if statements that dissapear when false easy to
construct.

Most of the newline parsing was stolen from the mustache spec. However,
we're not currently tokenizing other whitespace. So, tags that arent
directly surrounded by newlines (started in the middle of a line of
text, for instance), will be rendered verbatim.